fix: validate UI5 version for card generator with enhanced error message
