+++ packages/app-config-writer/src/index.ts
@@ -1,7 +1,7 @@
 export { getSmartLinksTargetFromPrompt, simulatePrompt, includeTestRunnersPrompt } from './prompt/index.js';
 export { generateSmartLinksConfig } from './smartlinks-config/index.js';
 export { generateEslintConfig, convertEslintConfig } from './eslint-config/index.js';
-export { generateInboundNavigationConfig, readManifest } from './navigation-config/index.js';
+export { generateInboundNavigationConfig } from './navigation-config/index.js';
 export { generateVariantsConfig } from './variants-config/index.js';
 export { convertToVirtualPreview } from './preview-config/index.js';
 export { enableCardGeneratorConfig } from './cards-config/index.js';
